% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castxml.R
\name{install_castxml}
\alias{install_castxml}
\alias{download_castxml}
\alias{locate_castxml}
\title{Manage CastXML installation}
\usage{
install_castxml(version = "latest", force = FALSE)

download_castxml(version = "latest", dir = tempdir())

locate_castxml(path = getOption("porter.castxml"), patch = FALSE)
}
\arguments{
\item{version}{A valid CastXML version. If \code{"latest"}, which is the default,
the latest version is downloaded.}

\item{force}{If \code{TRUE}, reinstall even there is already a CastXML
installation in the same folder. Default: \code{FALSE}.}

\item{dir}{A directory to save the downloaded CastXML pre-buit binaries. It
will be created if not exists. Default: \code{tempdir()}.}

\item{path}{A directory where the CastXML executable exists. Default is use
the option \code{porter.castxml}. If empty, it will use the first value of:
\itemize{
\item Option \code{porter.castxml}
\item Value of \code{dirname(Sys.which("castxml"))}
\item Default installation directory for \code{install_castxml()}
\itemize{
\item For Windows: \code{Sys.getenv("APPDATA")/castxml}
\item For macOS: \verb{~/Library/Application Support/castxml}
\item For Linux: \verb{~/.local/share/castxml}
}
}}

\item{patch}{Only works for CastXML v0.4.8 on Windows. The superbuild of that
version has a missing dependency of \code{zlib} library on Windows. If
\code{TRUE}, the \code{zlib} library will be downloaded into the CastXML
installation directory. Default: \code{FALSE}.}
}
\value{
\itemize{
\item \code{download_castxml()} returns the full path of the downloaded CastXML
binaries, with an attribute \code{version} being the actual downloaded version
of CastXML.
\item \code{install_castxml()} returns either \code{TRUE} or \code{FALSE}.
\item \code{locate_castxml()} returns a single named character vector with the name
being the CastXML version and value being the full path of CastXML
executable, or \code{NULL} if no CastXML is found.
}
}
\description{
\code{download_castxml()} and \code{install_castxml()} download and install CastXML via the
\href{https://github.com/CastXML/CastXMLSuperbuild}{CastXML Superbuild} project,
respectively. \code{locate_castxml()} finds the path of CastXML executable.
}
\details{
CastXML is a C-family abstract syntax tree XML output tool. It is maintained
by Kitware in support of ITK, the Insight Segmentation and Registration
Toolkit.

\code{install_castxml()} downloads the pre-built binaries of CastXML together with
the LLVM/Clang dependencies via the
\href{https://github.com/CastXML/CastXMLSuperbuild}{CastXML Superbuild} project.

The location of the installed CastXML is OS-specific:
\itemize{
\item For Windows: \code{Sys.getenv("APPDATA")}
\item For macOS: \verb{~/Library/Application Support}
\item For Linux: \verb{~/.local/share}
}
}
\note{
Instead of using \code{install_castxml()}, you may directly use your package
manager to install CastXML, as it may already provide a \code{castxml} package.
\itemize{
\item On Windows, you can use \href{https://scoop.sh/}{Scoop} to install CastXML via:
\verb{scoop install main/castxml}
\item On macOS, you can use \href{https://brew.sh/}{Homebrew} to install CastXML via:
\verb{brew install castxml}
\item On Linux, you can check if your system package manager provides CastXML or
not via \href{https://repology.org/project/castxml/versions}{Repology}.
}
}
\examples{
\dontrun{
download_castxml("0.5.0", tempdir())

install_castxml("latest")

locate_castxml()
}
}
