% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{port}
\alias{port}
\title{Parse C-family header files using CastXML and gather data from the generated
XML.}
\usage{
port(
  header,
  limit = TRUE,
  keep = FALSE,
  cflags = NULL,
  castxml = locate_castxml()
)
}
\arguments{
\item{header}{A single string indicating the path of C header file.}

\item{limit}{A flag (\code{TRUE}/\code{FALSE}) a string, or a character vector.
If \code{TRUE}, only functions/enums/structs/unions from the directory of
the input header are kept. If \code{FALSE}, no subsetting is performed.
If a single string, it will be used as a regular expression to extract
only functions/enums/structs/unions that match the pattern.
If a character vector, it should be paths of directories used to
constrain the scope.
Default: \code{TRUE}.}

\item{keep}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE}, the XML file generated by
CastXML is saved into the same directory of input header file with the
same name. Default: \code{FALSE}.}

\item{cflags}{A character vector of C compile flags passed to CastXML.
Default: \code{NULL}.}

\item{castxml}{A single string indicating the file path of the CastXML
executable. Default is used the path found by \code{\link[=locate_castxml]{locate_castxml()}}.}
}
\value{
An \code{dynport} object.
}
\description{
Parse C-family header files using CastXML and gather data from the generated
XML.
}
