% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{port_fields}
\alias{port_fields}
\alias{port_has}
\alias{port_get}
\alias{port_set}
\title{Query and update \code{dynport} fields}
\usage{
port_fields(port)

port_has(port, fields)

port_get(port, fields)

port_set(port, ...)
}
\arguments{
\item{port}{A \code{dynport} object created using \link{port}.}

\item{fields}{A character vector of valid field names in input \code{dynport}
object.}

\item{...}{Pairs of field names and values. See Details.}
}
\value{
\code{port_fields()} returns a character vector.

\code{port_has()} returns a logical vector.

\code{port_get()} returns a character vector or a data.frame, depending on the
field value types.

\code{port_set()} returns a new \code{dynport} object, invisibly.
}
\description{
Query and update \code{dynport} fields
}
\details{
\code{port_fields()} lists the fields in the object.

\code{port_has()} checks if one or multiple fields are present in the object.

\code{port_get()} returns the current values of input fields in the object. If
multiple fields are given, a named list is returned. Note that \code{port_get()}
errors if non-existing fields are queried.

\code{port_set()} updates one or multiple field values in the object. It accepts
two forms of arguments:
\itemize{
\item Two arguments with the first element being the name of a single field, and
the second being the field value. Use \code{NULL} value to remove the field.
\item Multiple named arguments with names being names of field to set, and values
being the new values. Use \code{NULL} value to remove the field.
}

For \code{port_set()}, you can use \code{NULL} to remove values of input fields. For
example, the following two lines of code both remove the value of the
\code{Version} field.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{port_set(p, "Version", NULL)
port_set(p, Version = NULL)
}\if{html}{\out{</div>}}

Note that, following fields are always required for a \code{dynport} object and
will be kept as \code{NULL} instead of removal: \code{Package}, \code{Version}, \code{Library},
\code{Function}, \code{FuncPtr}, \code{Enum}, \code{Struct} and \code{Union}. Other fields are treated
as user customize fields and will be removed when setting to \code{NULL}.`
}
