% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{port_write}
\alias{port_write}
\title{Write \code{dynport} file}
\usage{
port_write(port, file, reorder = TRUE)
}
\arguments{
\item{port}{A \code{dynport} object}

\item{file}{A single string or a connection object.}

\item{reorder}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE}, required \code{dynport} fields
are always written before all other customized fields. Default:
\code{TRUE}.}
}
\value{
A single string of the file path, invisibly.
}
\description{
Write \code{dynport} file
}
\examples{
\dontshow{if (!is.null(porter::locate_castxml())) withAutoprint(\{ # examplesIf}
header <- tempfile(fileext = ".h")
writeLines("int add(int a, int b);", header)

p <- port(header)
p <- port_set(p, Package = "Example", Version = "1.0", Library = "example")
port_write(p, file.path(tempdir(), "Example.dynport"))
\dontshow{\}) # examplesIf}
}
