% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{port_xml}
\alias{port_xml}
\title{Extract data from the XML file generated by CastXML}
\usage{
port_xml(xml, dirs = NULL)
}
\arguments{
\item{xml}{A path of an XML generated using CastXML. Both the
\code{--castxml-output} and \code{--output-gccxml} options are supported.}

\item{dirs}{A character of directories used to constrain the header files to
extract the data. Default: \code{NULL}.}
}
\value{
A named list of six elements:
\itemize{
\item \code{func}: function data
\item \code{enum}: enumeration data
\item \code{struct}: struct data
\item \code{union}: union data
\item \code{type}: type data
\item \code{file}: file data
}
}
\description{
Extract data from the XML file generated by CastXML
}
\keyword{internal}
