% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porter-package.R
\docType{package}
\name{porter-package}
\alias{porter}
\alias{porter-package}
\title{porter: Generate Port Files for C Libraries}
\description{
Generate port files for foreign function interfaces to C libraries by parsing C-family header files with 'CastXML'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hongyuanjia/porter}
  \item Report bugs at \url{https://github.com/hongyuanjia/porter/issues}
}

}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqust.edu.cn} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID}) [copyright holder]

}
\keyword{internal}
