% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\name{print.dynport}
\alias{print.dynport}
\title{Print summary of an \code{dynport} object}
\usage{
\method{print}{dynport}(x, n = 5L, ...)
}
\arguments{
\item{x}{A \code{dynport}.}

\item{n}{The number of items to print. Default: \code{5}.}

\item{...}{Other arguments to pass. Reserved for future use.}
}
\value{
\code{x} invisibly.
}
\description{
Print summary of an \code{dynport} object
}
\examples{
\dontshow{if (!is.null(porter::locate_castxml())) withAutoprint(\{ # examplesIf}
header <- tempfile(fileext = ".h")
writeLines("int add(int a, int b);", header)

p <- port(header)
print(p)
\dontshow{\}) # examplesIf}
}
