% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{port_write}
\alias{port_write}
\title{Write \code{dynport} file}
\usage{
port_write(port, file, reorder = TRUE)
}
\arguments{
\item{port}{A \code{dynport} object}

\item{file}{A single string or a connection object.}

\item{reorder}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE}, required \code{dynport} fields
are always written before all other customized fields. Default:
\code{TRUE}.}
}
\value{
A single string of the file path, invisibly.
}
\description{
Write \code{dynport} file
}
