% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porter-package.R
\docType{package}
\name{porter-package}
\alias{porter}
\alias{porter-package}
\title{porter: Generate Port Files for Foreign Function Interface for C Libraries}
\description{
Generate port files for parsing header files of C libraries using 'CastXML'.
}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqust.edu.cn} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID})

}
\keyword{internal}
