% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porter-package.R
\docType{package}
\name{porter-package}
\alias{porter}
\alias{porter-package}
\title{porter: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuanjia@cqust.edu.cn} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID})

}
\keyword{internal}
